moved test into a folder and splitted arduino code and raspberry pi code
