+++ tests/CMakeLists.txt
@@ -33,7 +33,7 @@ target_link_libraries(main_test PUBLIC
         data sensors telecom ERT_RF_Protocol_Interface
         adxl375 bmi08a bmi08g bmi08xa bmp3 I2CGPS TinyGPS++)
 
-set (I2CARDUINO_TEST src/I2CArduino_test.cpp)
+set (I2CARDUINO_TEST src/I2CArduino_test/I2CArduino_RaspberryPi.cpp)
 add_executable(i2c_arduino_test ${I2CARDUINO_TEST})
 target_link_libraries(i2c_arduino_test PUBLIC I2CInterface)
